qcacld-3.0: Flow Policy Manager(FPM) (nl80211 User Interface)
a. Exposes NL interface to user to configure flows.
b. Allow user to add/delete/update policies.
c. Remove cookie and use only policy_id as 64-bit
   unique identifier.

CRs-Fixed: 3692806

+++ components/dp/dispatcher/inc/wlan_dp_ucfg_api.h
@@ -1700,7 +1700,7 @@ QDF_STATUS ucfg_fpm_policy_update(struct fpm_table *fpm,
  *
  * Return: 0 if FPM policy removed successfully
  */
-QDF_STATUS ucfg_fpm_policy_rem(struct fpm_table *fpm, uint32_t cookie);
+QDF_STATUS ucfg_fpm_policy_rem(struct fpm_table *fpm, uint64_t cookie);
 
 /*
  * ucfg_fpm_policy_get() - Get FPM policy array
@@ -1770,7 +1770,7 @@ QDF_STATUS ucfg_fpm_policy_update(struct fpm_table *fpm,
 }
 
 static inline
-QDF_STATUS ucfg_fpm_policy_rem(struct fpm_table *fpm, uint32_t cookie)
+QDF_STATUS ucfg_fpm_policy_rem(struct fpm_table *fpm, uint64_t cookie)
 {
 	return QDF_STATUS_SUCCESS;
 }
